refactor: extract PaneContext from SikulixIDE inner class to top-level class
Move the 656-line PaneContext inner class to its own file (PaneContext.java).
All implicit references to the enclosing SikulixIDE instance are replaced
with explicit SikulixIDE.get() calls. External references in EditorPane,
ButtonCapture, PatternWindow, and SikuliIDEPopUpMenu updated from
SikulixIDE.PaneContext to PaneContext.

SikulixIDE.java reduced from ~3319 to ~2665 lines.
